BaseDir/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2
djangorestframework==3.14.0
exponent-server-sdk==2.0.0
idna==3.4
Pillow==10.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.4
whitenoise==6.5.0
